+++ azure-dls-gen2/src/main/java/org/talend/components/adlsgen2/service/BlobInformations.java
@@ -39,5 +39,5 @@ public class BlobInformations {
 
     private String directory;
 
-    private Integer contentLength = 0;
+    private Long contentLength = 0L;
 }
